feat: removed header, footer, and menu links
