--- apps/web/src/components/MenuLinks/__tests__/index.test.tsx
@@ -1,14 +1,0 @@
-import React from 'react';
-import { composeStories } from '@storybook/testing-react';
-import { render } from '@testing-library/react';
-
-import projectAnnotations from 'ui/utils/testSetup';
-import * as stories from '../MenuLinks.stories';
-
-const { Defaults } = composeStories(stories, projectAnnotations);
-
-describe('MenuLinks', () => {
-  it('component mounts', () => {
-    render(<Defaults />);
-  });
-});
