+++ apps/web/src/components/index.tsx
@@ -1,10 +1,6 @@
 import dynamic from 'next/dynamic';
 
-export const ImageWithContent = dynamic(() => import('./ImageWithContent'), {
-  ssr: true
-}) as React.FC;
+export const ImageWithContent = dynamic(() => import('./ImageWithContent')) as React.FC;
 
 // CODE INJECTION COMMENT
-export const Hero = dynamic(() => import('./Hero'), {
-  ssr: true
-}) as React.FC;
+export const Hero = dynamic(() => import('./Hero')) as React.FC;
